🌈 style(studio): Major changes
- Added style variables to globals.css
- Repositioned page routes
- Added favicon
- Renamed `cubicsui-studio` to only `studio`
