--- apps/cubicsui-studio/app/dashboard/page.tsx
@@ -1,3 +1,0 @@
-export default function DashboardPage() {
-  return <div>This is the dashboard</div>;
-}
