--- apps/cubicsui-studio/app/layout.tsx
@@ -1,25 +1,0 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-
-const inter = Inter({
-  variable: "--font-family",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "CubicsUI Studio",
-  description: "Compose and manage components using CubicsUI Studio",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={`${inter.variable}`}>{children}</body>
-    </html>
-  );
-}
