--- apps/cubicsui-studio/app/page.tsx
@@ -1,9 +1,0 @@
-import Link from "next/link";
-
-export default async function Home() {
-  return (
-    <div>
-      <Link href={"/dashboard"}>Dashboard</Link>
-    </div>
-  );
-}
