+++ apps/studio/.gitignore
