+++ apps/studio/app/components/[id]/page.tsx
