+++ apps/studio/app/components/page.module.css
@@ -1,0 +1,4 @@
+.container {
+  display: flex;
+  flex-direction: column;
+}
