+++ apps/studio/app/globals.css
@@ -1,0 +1,108 @@
+:root {
+  /* Color Pallete */
+  /* gray */
+  --gray-100: rgb(255, 255, 255);
+  --gray-75: rgb(191, 191, 191);
+  --gray-50: rgb(128, 128, 128);
+  --gray-25: rgb(64, 64, 64);
+  --gray-0: rgb(23, 23, 23);
+
+  /* reds */
+  --red-100: rgb(255, 105, 92);
+  --red-75: rgb(255, 128, 128);
+  --red-50: rgb(255, 163, 163);
+  --red-25: rgb(255, 198, 198);
+
+  /* greens */
+  --green-100: rgb(0, 255, 128);
+  --green-75: rgb(64, 255, 160);
+  --green-50: rgb(128, 255, 192);
+  --green-25: rgb(192, 255, 224);
+
+  /* blues */
+  --blue-100: rgb(68, 68, 255);
+  --blue-75: rgb(110, 110, 255);
+  --blue-50: rgb(151, 151, 255);
+  --blue-25: rgb(193, 193, 255);
+
+  /* yellows */
+  --yellow-100: rgb(220, 255, 66);
+  --yellow-75: rgb(229, 255, 112);
+  --yellow-50: rgb(236, 255, 153);
+  --yellow-25: rgb(244, 255, 194);
+
+  /* oranges */
+  --orange-100: rgb(255, 173, 66);
+  --orange-75: rgb(255, 193, 112);
+  --orange-50: rgb(255, 211, 153);
+  --orange-25: rgb(255, 228, 194);
+
+  /* Base colors */
+  --background: var(--gray-100);
+  --foreground: var(--gray-0);
+  --primary: var(--blue-100);
+  --success: var(--green-100);
+  --danger: var(--red-100);
+  --warning: var(--orange-100);
+
+  /* Derived colors */
+
+  /*This will be used for any surface that needs to be separated from the background*/
+  --overlay: color-mix(in hsl, var(--background) 90%, var(--gray-50));
+
+  --text: var(--foreground);
+  --hl-text: var(--primary);
+  --text-faint: color-mix(in srgb, var(--text) 50%, transparent);
+
+  --primary-faint: color-mix(in hsl, var(--primary) 20%, transparent);
+  --border-color: color-mix(in hsl, var(--foreground) 20%, transparent);
+  --shadow-color: color-mix(in hsl, var(--foreground) 30%, transparent);
+
+  /* Spacings */
+  --sp-1: 8px;
+  --sp-2: 12px;
+  --sp-3: 16px;
+  --sp-4: 24px;
+  --sp-5: 32px;
+  --sp-6: 48px;
+
+  /* Utilities */
+  --br: 0.5em;
+  --border: 1px solid var(--border-color);
+  --shadow: 0 0 10px var(--shadow-color);
+  --transition-tf: cubic-bezier(0.785, 0.135, 0.15, 0.86);
+  --header-height: 4rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    /* Base colors */
+    --background: var(--gray-0);
+    --foreground: var(--gray-100);
+  }
+}
+
+body {
+  color: var(--text);
+  background: var(--background);
+  font-family: var(--font-family), Arial, Helvetica, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+@media (prefers-color-scheme: dark) {
+  html {
+    color-scheme: dark;
+  }
+}
