+++ apps/studio/app/page.tsx
@@ -1,0 +1,4 @@
+
+export default async function Home() {
+  return <div>This is Home</div>;
+}
