+++ apps/studio/package.json
